# api/src/main/resources/openapi/admin-api.yaml
openapi: 3.0.3
info:
  title: Filocitas Admin API
  version: 1.0.0
  description: Administrative API for managing quotes, authors, and service health for Filocitas.
servers:
  - url: /api
paths:
  /admin/quotes:
    post:
      tags:
        - Admin
      summary: Create multiple quotes
      operationId: createQuotes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/QuoteRequest'
      responses:
        '201':
          description: Quotes created successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: 'public-api.yaml#/components/schemas/Quote'
  /admin/quote:
    post:
      tags:
        - Admin
      summary: Create a quote
      operationId: createQuote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
      responses:
        '201':
          description: Quote created successfully
          content:
            application/json:
              schema:
                $ref: 'public-api.yaml#/components/schemas/Quote'
  /admin/authors/quotes/limit/{limit}:
    get:
      tags:
        - Admin
      summary: Get the top authors with their quote counts
      description: Returns the authors with the most quotes ordered in descending order, limited to the requested amount.
      operationId: getAuthorsQuotesCountLimit
      parameters:
        - name: limit
          in: path
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        '200':
          description: Top authors with quote counts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthorQuotesCount'
        '204':
          description: No authors found
        '400':
          description: Invalid limit provided
  /admin/export:
    get:
      tags:
        - Admin
      summary: Export all quotes in a seed-friendly format
      description: Returns the complete collection of quotes using the same structure as the creation payload so it can be re-imported if needed.
      operationId: exportQuotes
      responses:
        '200':
          description: Quotes exported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteExportResponse'
  /admin/quote/{id}:
    delete:
      tags:
        - Admin
      summary: Delete a quote
      operationId: deleteQuote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Quote deleted successfully
  /admin/ping:
    get:
      tags:
        - Health
      summary: Ping the service
      operationId: ping
      responses:
        '200':
          description: Pong message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PingResponse'
components:
  schemas:
    QuoteRequest:
      type: object
      required:
        - quote
        - author
      properties:
        quote:
          type: string
          description: The textual content of the quote.
        author:
          type: string
          description: The display name of the quote author.
        author_wikipedia:
          type: string
          format: uri
          description: Optional link to the author's Wikipedia page.
        work:
          type: string
          description: The work where the quote originated.
        year:
          type: integer
          format: int32
          description: The year the quote was created or published. Negative numbers represent BCE years.
        tags:
          type: array
          description: Thematic tags associated with the quote.
          items:
            type: string
        century:
          type: string
          description: The century associated with the quote, e.g., "-5".
    PingResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    AuthorQuotesCount:
      type: object
      required:
        - author
        - quotes_count
      properties:
        author:
          $ref: 'public-api.yaml#/components/schemas/Author'
        quotes_count:
          type: integer
          format: int64
          minimum: 0
    
    QuoteExportResponse:
      type: object
      required:
        - quotes
      properties:
        quotes:
          type: array
          description: Exported quotes ready to be re-imported as creation payloads.
          items:
            $ref: '#/components/schemas/QuoteRequest'